finance salaire et accompte

+++ src/Controller/AccompteController.php
@@ -92,23 +92,27 @@ final class AccompteController extends AbstractController
         if ($form->isSubmitted() && $form->isValid()) {
             $accomp= $entityManager->getRepository(Accompte::class)->findOneBy(['employe' => $accompte->getEmploye()->getId(), 'paye' => false], ['id' =>'DESC']);
             if(empty($accomp)){
-            $entityManager->persist($accompte);
-            $entityManager->flush();
+                if($accompte->getMontant() <= $accompte->getEmploye()->getPoste()->getSalaire()){
+                    $entityManager->persist($accompte);
+                    $entityManager->flush();
 
-            $this->addFlash('notice', 'enrgistre avec succes');
-            $response = $this->redirectToRoute('app_accompte_index', [], Response::HTTP_SEE_OTHER);
-                $response->setSharedMaxAge(0);
-                $response->headers->addCacheControlDirective('no-cache', true);
-                $response->headers->addCacheControlDirective('no-store', true);
-                $response->headers->addCacheControlDirective('must-revalidate', true);
-                $response->setCache([
-                    'max_age' => 0,
-                    'private' => true,
-                ]);
-                return $response;
+                    $this->addFlash('notice', 'enrgistre avec succes');
+                    $response = $this->redirectToRoute('app_accompte_index', [], Response::HTTP_SEE_OTHER);
+                        $response->setSharedMaxAge(0);
+                        $response->headers->addCacheControlDirective('no-cache', true);
+                        $response->headers->addCacheControlDirective('no-store', true);
+                        $response->headers->addCacheControlDirective('must-revalidate', true);
+                        $response->setCache([
+                            'max_age' => 0,
+                            'private' => true,
+                        ]);
+                        return $response;
                 }else{
-                    $this->addFlash('notice', 'l\'employé a déjà un accompte non remboursé');
-                }
+                        $this->addFlash('notice', 'l\'accompte ne peut etre suprerieur au salaire de base');
+                    }
+            }else{
+            $this->addFlash('notice', 'l\'employé a déjà un accompte non remboursé');
+             }
         }
 
        
@@ -173,22 +177,41 @@ final class AccompteController extends AbstractController
     #[Route('/{id}/edit', name: 'app_accompte_edit', methods: ['GET', 'POST'])]
     public function edit(Request $request, Accompte $accompte, EntityManagerInterface $entityManager): Response
     {
-        if ($this->security->isGranted('ROLE_RH')) {
-        $form = $this->createForm(AccompteForm::class, $accompte);
-        $form->remove('employe');
-        $form->handleRequest($request);
+        if ($this->security->isGranted('ROLE_RH')) {// && !$accompte->getVerser()
+           
 
-        if ($form->isSubmitted() && $form->isValid()) {
-            $entityManager->flush();
+                $form = $this->createForm(AccompteForm::class, $accompte);
+                $form->remove('employe');
+                $form->handleRequest($request);
 
-            return $this->redirectToRoute('app_accompte_index', [], Response::HTTP_SEE_OTHER);
-        }
+                if ($form->isSubmitted() && $form->isValid() && !$accompte->getVerser()) {
+                    if (!$accompte->getVerser()) {
+                        if($accompte->getMontant() <= $accompte->getEmploye()->getPoste()->getSalaire()){
+                        $entityManager->flush();
+
+                       
+                        $response = $this->redirectToRoute('app_accompte_index', [], Response::HTTP_SEE_OTHER);
+                        $response->setSharedMaxAge(0);
+                        $response->headers->addCacheControlDirective('no-cache', true);
+                        $response->headers->addCacheControlDirective('no-store', true);
+                        $response->headers->addCacheControlDirective('must-revalidate', true);
+                        $response->setCache([
+                            'max_age' => 0,
+                            'private' => true,
+                        ]);
+                        return $response;
+                    }else{
+                        $this->addFlash('notice', 'l\'accompte ne peut etre suprerieur au salaire de base');
+                    }
+                    }else  $this->addFlash('notice', 'Accompte deja verse');
+                }
+            
 
        
-        $response = $this->render('accompte/edit.html.twig', [
-            'accompte' => $accompte,
-            'form' => $form,
-        ]);
+            $response = $this->render('accompte/edit.html.twig', [
+                'accompte' => $accompte,
+                'form' => $form,
+            ]);
             $response->setSharedMaxAge(0);
             $response->headers->addCacheControlDirective('no-cache', true);
             $response->headers->addCacheControlDirective('no-store', true);
@@ -216,10 +239,11 @@ final class AccompteController extends AbstractController
     public function delete(Request $request, Accompte $accompte, EntityManagerInterface $entityManager): Response
     {
         if ($this->security->isGranted('ROLE_RH')) {
-        if ($this->isCsrfTokenValid('delete'.$accompte->getId(), $request->getPayload()->getString('_token'))) {
+        if ($this->isCsrfTokenValid('delete'.$accompte->getId(), $request->getPayload()->getString('_token')) && !$accompte->getVerser()) {
             $entityManager->remove($accompte);
             $entityManager->flush();
-        }
+       
+        }else $this->addFlash('notice', 'Accompte deja verse');
 
         
         $response = $this->redirectToRoute('app_accompte_index', [], Response::HTTP_SEE_OTHER);
